pyproject.toml
[project]
name = "HQLLM"
version = "0.1.0"
description = "Streamlit-based mini RAG bot that pulls data from Google Drive or through file upload"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit==1.38.0",
    "langchain==0.3.0",
    "langchain-anthropic==0.2.1",
    "langchain-chroma==0.1.4",
    "langchain-community==0.3.0",
    "langchain-core==0.3.1",
    "langchain-openai==0.2.0",
    "langchain-text-splitters==0.3.0",
    "langsmith==0.1.123",
    "pydantic==2.7.4",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "sentence-transformers==2.3.0",
    "google-api-python-client==2.108.0",
    "google-auth-httplib2==0.1.1",
    "google-auth-oauthlib==1.1.0",
    "chromadb==0.4.15",
    "python-dotenv==1.0.0",
    "ruff==0.1.9",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.ruff]
line-length = 100
target-version = "py311"
